requirements.txt
annotated-types==0.7.0
anyio==4.10.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.1.7
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.112.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Login==0.6.3
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
gevent==25.5.1
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
janus==1.0.0
Jinja2==3.1.6
kafka==1.3.5
kafka-python==2.2.15
MarkupSafe==3.0.2
numpy==2.0.1
openpyxl==3.1.5
packaging==25.0
pandas==2.2.2
phpserialize==1.3
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2025.2
requests==2.32.3
simple-websocket==1.1.0
simpy==4.1.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.37.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.0.3
wsproto==1.2.0
WTForms==3.1.2
zope.event==5.1.1
zope.interface==7.2
